@@ -1,6 +1,20 @@
 import React from "react";
-import HelloWorld from "components/HelloWorld";
+import {BrowserRouter as Router, Routes, Route} from "react-router-dom";
 
-const App: React.FC = () => <HelloWorld />;
+import AppPage from "./pages/AppPage";
+import FAQPage from "./pages/FAQPage";
 
-export default App;
+import "./styling/app.css";
+
+export default function App() {
+    return (
+        <>
+            <Router>
+                <Routes>
+                    <Route path="/" element={<AppPage />} />
+                    <Route path="/faq" element={<FAQPage />} />
+                </Routes>
+            </Router>
+        </>
+    );
+}
